+++ CHANGELOG.md
@@ -5,6 +5,10 @@ they can and will change without that change being reflected in Styler's semanti
 
 ## main
 
+### Fixes
+
+- also leave piped versions of the shift + microsecond bug alone (`|> DateTime.add(val, :microsecond)` (#256, h/t @zvlasic)
+
 ## 1.12.1
 
 ### Fixes